Replace node/cli with node/cli2
