+++ node/cli/src/lib.rs
@@ -1,10 +1,13 @@
 extern crate beacon;
+extern crate beacon_chain_handler;
+extern crate chain;
 extern crate clap;
 extern crate client;
+extern crate futures;
 extern crate network;
+extern crate node_rpc;
 extern crate node_runtime;
 extern crate primitives;
-extern crate parking_lot;
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
@@ -12,55 +15,96 @@ extern crate serde_derive;
 extern crate serde_json;
 extern crate service;
 extern crate storage;
-#[macro_use]
-extern crate log;
+extern crate tokio;
+extern crate parking_lot;
 
-use beacon::types::BeaconBlockHeader;
-use chain_spec::{deserialize_chain_spec, get_default_chain_spec};
+use beacon::types::{BeaconBlock, BeaconBlockHeader};
+use chain::BlockChain;
 use clap::{App, Arg};
 use client::Client;
+use futures::future;
+use futures::sync::mpsc::channel;
 use network::protocol::ProtocolConfig;
 use network::service::{
     generate_service_task, NetworkConfiguration, Service as NetworkService
 };
+use node_rpc::api::RpcImpl;
+use node_runtime::{Runtime, StateDbViewer};
+use primitives::hash::CryptoHash;
 use primitives::signer::InMemorySigner;
-use primitives::traits::{GenericResult, Signer};
-use network::network_handler::NetworkHandler;
-use service::run_service;
-use std::fs::File;
-use std::io::Read;
-use std::path::{Path, PathBuf};
+use network::network_handler::ChannelNetworkHandler;
+use std::path::Path;
 use std::sync::Arc;
-use std::time::Duration;
+use storage::{StateDb, Storage};
 use parking_lot::RwLock;
 
 pub mod chain_spec;
 
-fn get_storage_path(base_path: &Path) -> PathBuf {
-    let mut path = base_path.to_owned();
-    path.push("storage/db");
-    path
+fn get_storage(base_path: &Path) -> Arc<Storage> {
+    let mut storage_path = base_path.to_owned();
+    storage_path.push("storage/db");
+    Arc::new(storage::open_database(&storage_path.to_string_lossy()))
 }
 
-fn start_service(base_path: &Path, chain_spec_path: Option<&Path>) -> GenericResult {
-    let chain_spec = match chain_spec_path {
-        Some(path) => {
-            let mut file = File::open(path).expect("could not open chain spec file");
+fn start_service(base_path: &Path, chain_spec_path: Option<&Path>) {
+    // Create shared-state objects.
+    let storage = get_storage(base_path);
+    let chain_spec = chain_spec::read_or_default_chain_spec(&chain_spec_path);
 
-            let mut contents = String::new();
-            file.read_to_string(&mut contents).expect("could not read from chain spec file");
+    let state_db = Arc::new(StateDb::new(storage.clone()));
+    let runtime = Arc::new(RwLock::new(Runtime::new(state_db.clone())));
+    let genesis_root = runtime.write().apply_genesis_state(
+        &chain_spec.accounts,
+        &chain_spec.genesis_wasm,
+        &chain_spec.initial_authorities,
+    );
 
-            deserialize_chain_spec(&contents)
-        }
-        None => get_default_chain_spec(),
-    }.unwrap();
+    let genesis = BeaconBlock::new(
+        0, CryptoHash::default(), genesis_root, vec![], vec![]
+    );
+    let beacon_chain = Arc::new(BlockChain::new(genesis, storage.clone()));
 
-    let signer = Arc::new(InMemorySigner::new());
-    info!("Public key: {}", signer.public_key());
-    let storage_path = get_storage_path(base_path);
-    let storage = Arc::new(storage::open_database(&storage_path.to_string_lossy()));
+    // Create RPC Server.
+    let state_db_viewer = StateDbViewer::new(beacon_chain.clone(), state_db.clone());
+    // TODO: TxFlow should be listening on these transactions.
+    let (submit_txn_tx, _submit_txn_rx) = channel(1024);
+    let (submit_receipt_tx, _submit_receipt_rx) = channel(1024);
+    let rpc_impl = RpcImpl::new(state_db_viewer, submit_txn_tx.clone());
+    let rpc_handler = node_rpc::api::get_handler(rpc_impl);
+    let server = node_rpc::server::get_server(rpc_handler);
+    tokio::spawn(future::lazy(|| {
+        server.wait();
+        Ok(())
+    }));
+
+    // Create a task that consumes the consensuses and produces the beacon chain blocks.
+    let signer = Arc::new(InMemorySigner::default());
+    let (
+        _beacon_block_consensus_body_tx,
+        beacon_block_consensus_body_rx,
+    ) = channel(1024);
+    let block_producer_task = beacon_chain_handler::producer::create_beacon_block_producer_task(
+        beacon_chain.clone(),
+        runtime.clone(),
+        signer.clone(),
+        state_db.clone(),
+        beacon_block_consensus_body_rx,
+    );
+    tokio::spawn(block_producer_task);
+
+    // Create task that can import beacon chain blocks from other peers.
+    let (_beacon_block_tx, beacon_block_rx) = channel(1024);
+    let block_importer_task = beacon_chain_handler::importer::create_beacon_block_importer_task(
+        beacon_chain.clone(),
+        runtime.clone(),
+        state_db.clone(),
+        beacon_block_rx
+    );
+    tokio::spawn(block_importer_task);
+
+    // Create network task.
+    let network_handler = ChannelNetworkHandler::new(submit_txn_tx.clone(), submit_receipt_tx.clone());
     let client = Arc::new(RwLock::new(Client::new(&chain_spec, storage, signer)));
-    let network_handler = NetworkHandler { client: client.clone() };
     let network = NetworkService::new(
         ProtocolConfig::default(),
         NetworkConfiguration::default(),
@@ -71,8 +115,8 @@ fn start_service(base_path: &Path, chain_spec_path: Option<&Path>) -> GenericRes
         &network.network,
         &network.protocol,
     );
-    let produce_blocks_interval = Duration::from_secs(2);
-    run_service(&client, network_task, produce_blocks_interval)
+
+    tokio::run(network_task);
 }
 
 pub fn run() {
@@ -98,5 +142,5 @@ pub fn run() {
 
     let chain_spec_path = matches.value_of("chain_spec_file").map(|x| Path::new(x));
 
-    start_service(base_path, chain_spec_path).unwrap();
+    start_service(base_path, chain_spec_path);
 }
