--- node/cli2/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "node-cli2"
-version = "0.1.0"
-
-[dependencies]
-clap = "2.32.0"
-futures = "0.1.25"
-tokio = "0.1.11"
-
-beacon = { path = "../../core/beacon" }
-beacon-chain-handler = { path = "../beacon-chain-handler" }
-chain = { path = "../../core/chain" }
-client = { path = "../client" }
-network = { path = "../network", features = ["test-utils"] }
-node-runtime = { path = "../runtime" }
-node-rpc = { path = "../rpc" }
-parking_lot = "0.6"
-primitives = { path = "../../core/primitives" }
-storage = { path = "../../core/storage" }
-service = { path = "../service" }
-
-serde = "1.0"
-serde_derive = "1.0"
-serde_json = "1.0"
